+++ src/detections/detection.rs
@@ -284,10 +284,16 @@ impl Detection {
                         key.as_str(),
                         Channel(
                             stored_static
-                                .ch_config
-                                .get(&CompactString::from(ch_str.to_ascii_lowercase()))
-                                .unwrap_or(ch_str)
-                                .to_owned(),
+                                .ch_disp_abbr_generic
+                                .replace_all(
+                                    stored_static
+                                        .ch_config
+                                        .get(&CompactString::from(ch_str.to_ascii_lowercase()))
+                                        .unwrap_or(ch_str)
+                                        .as_str(),
+                                    &stored_static.ch_disp_abbr_gen_rep_values,
+                                )
+                                .into(),
                         ),
                     );
                 }