+++ src/main.rs
@@ -1105,12 +1105,7 @@ impl App {
             ));
 
             // timeline機能の実行
-            tl.start(
-                &records_per_detect,
-                stored_static.metrics_flag,
-                stored_static.logon_summary_flag,
-                &stored_static.eventkey_alias,
-            );
+            tl.start(&records_per_detect, stored_static);
 
             if !(stored_static.metrics_flag || stored_static.logon_summary_flag) {
                 // ruleファイルの検知
@@ -1251,12 +1246,7 @@ impl App {
             ));
 
             // timeline機能の実行
-            tl.start(
-                &records_per_detect,
-                stored_static.metrics_flag,
-                stored_static.logon_summary_flag,
-                &stored_static.eventkey_alias,
-            );
+            tl.start(&records_per_detect, stored_static);
 
             if !(stored_static.metrics_flag || stored_static.logon_summary_flag) {
                 // ruleファイルの検知